refactor paths
