+++ code/spikes/deno.000/src.api.deno/common/libs.ts
