+++ code/spikes/deno.000/src.api.deno/mod.ts
