+++ code/spikes/deno.000/src.api.deno/types.ts
@@ -1,4 +1,6 @@
-export type * from '../src/t.deno.ts';
+// export type * from '../src/t.deno.ts';
+export type * from '../../../ext/ext.lib.deno/src/t.deno.ts';
+
 export type * from './route.subhosting/t.ts';
 
 /**
