+++ code/spikes/deno.000/src/main.ts
@@ -1,7 +1,8 @@
 // deno-lint-ignore-file no-explicit-any
 import { Server, EnvVars } from './common/mod.ts';
 import openai from './r.openai/mod.ts';
-import { Routes as deno } from '../../../ext/ext.lib.deno/src.deno/mod.ts';
+// import { Routes as deno } from '../../../ext/ext.lib.deno/src.deno/mod.ts';
+import { Routes as deno } from '../src.api.deno/mod.ts';
 
 /**
  * Initialize a new HTTP server.
