+++ code/spikes/deno.000/src/types.ts
@@ -1,2 +1,2 @@
-export type { HonoApp } from '../../../ext/ext.lib.deno/src.deno/types.ts';
+export type { HonoApp } from '../src.api.deno/types.ts';
 export type * from './r.openai/t.ts';
